requirements.txt
streamlit
langchain
python-dotenv
langchain-google-genai
google-api-python-client
Pillow
text2emotion
emoji==1.7.0